[241010] 이찬우  DAY28 과제 제출
