+++ my-app/src/components/boards-list/banner/styles.module.css
@@ -1,0 +1,10 @@
+.asd {
+  width: 100%;
+  margin-bottom: 50px;
+}
+
+.banner {
+  width: 100%;
+  height: 494px;
+  object-fit: cover;
+}
