+++ my-app/src/components/boards-list/list/queries.ts
@@ -1,0 +1,18 @@
+import { gql } from "@apollo/client";
+
+export const FETCH_BOARDS = gql`
+  query {
+    fetchBoards(page: 1) {
+      title
+      writer
+      createdAt
+      _id
+    }
+  }
+`;
+
+export const DELETE_BOARD = gql`
+  mutation deleteBoard($boardId: ID!) {
+    deleteBoard(boardId: $boardId)
+  }
+`;
