+++ my-app/src/components/boards-list/list/types.ts
@@ -1,0 +1,3 @@
+export interface IBoardProps {
+  styles: Record<string, string>;
+}
