--- my-app/src/components/boards-list/styles.module.css
@@ -1,95 +1,0 @@
-.boardContainer {
-  display: flex;
-  flex-direction: column;
-  min-width: 1000px;
-  padding: 24px 40px;
-  gap: 24px;
-  border-radius: 16px;
-  background: #fff;
-  box-shadow: 0px 0px 20px 0px rgba(0, 0, 0, 0.08);
-}
-
-.boardTop {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  padding: 16px 24px;
-}
-
-.boardTopLeft {
-  display: flex;
-  align-items: center;
-  gap: 50px;
-  font-weight: 500;
-  line-height: 20px; /* 125% */
-}
-
-.boardTopRight {
-  display: flex;
-  align-items: center;
-  gap: 50px;
-  font-weight: 500;
-  line-height: 20px; /* 125% */
-  padding-right: 27px;
-}
-
-.Author {
-  margin-right: 27px;
-}
-
-.boardContents {
-  display: flex;
-  justify-content: space-between;
-  padding: 12px 24px;
-  gap: 8px;
-
-  border-radius: 8px;
-  border: 1px solid var(--gray-50, #f2f2f2);
-  position: relative;
-}
-
-.boardNum {
-  overflow: hidden;
-  color: var(--Gray-50, var(--gray-500, #919191));
-  font-size: 14px;
-  font-weight: 300;
-}
-
-.boardTitle {
-  color: var(--Gray-90, var(--gray-900, #1c1c1c));
-  font-size: 14px;
-  font-weight: 500;
-}
-
-.boardSide {
-  display: flex;
-  align-items: center;
-  gap: 50px;
-  padding-right: 5px;
-}
-
-.boardName {
-  color: #333;
-  font-size: 14px;
-  font-weight: 300;
-  padding-right: 8px;
-}
-
-.boardDate {
-  color: var(--Gray-50, var(--gray-500, #919191));
-  font-size: 14px;
-  font-weight: 300;
-}
-
-.deleteIcon {
-  width: 24px;
-  height: 24px;
-  position: absolute;
-  right: 0;
-  opacity: 0; /* 처음에 보이지 않음 */
-  transition: opacity 0.3s ease; /* 부드러운 전환 효과 */
-}
-
-.boardContents:hover .deleteIcon {
-  opacity: 1; /* 호버 시 아이콘이 나타남 */
-}
