--- my-app/src/components/boards-list/types.ts
@@ -1,3 +1,0 @@
-export interface IBoardProps {
-  styles: Record<string, string>;
-}
